requirements.txt
Babel==2.2.0
dj-database-url==0.4.0
dj-static==0.0.6
Django==1.9.2
django-phonenumber-field==1.0.0
django-toolbelt==0.0.1
gunicorn==19.4.5
phonenumbers==7.2.5
psycopg2==2.6.1
pytz==2015.7
static3==0.6.1
wheel==0.26.0
